Life flight testing: Nowhere near ready, and event-posting/receiving is broken somehow.
Detail: Events that were expected to be crucial seem not to be posted by the 'bot,
so the landing/flying/takingoff events etcetera are not what Bebop seems to actually
use to indicate current behaviour. Fine. I reinstated the manual "flying" event sent
in the TakeOff method, and that event fires as expected... but landing isn't working
anymore, either in response to the event (??) or manually when issued after a delay.

Also, it seems as soon as I fix errors in telemetry readings, they stop sending
telemetry entirely. Somewhere in the pre-event-cue code, I think. Some events are
so spammy they have to be ignored, like CameraState and flattrim, but others
just don't seem to send at all.

Very odd. Needs lots more work. Also, a crucial next target for implementation
is ARDrone3:1:X set, because these seem to be the actual flight state telemetry..
though they don't get sent consistently, and therefore may not be reliable for
determining actual drone state in real time.